Restrangere derivate, eliminare dynamic_cast (o parte) si corectare greseli Am schimbat, in special, structura gandita pentru sageti si perk-uri. Perk-urile functioneaza acum printr-o singura clasa cu un atribut functie, iar ierarhia clasei sageata este: Sageata(Abstracta) -> SageataAimbot(calculul traiectoriei total diferit de restul), SageataRandom(damage, culoare si efect random), SageataDeBaza(doar da damage target-ului) -> SageataCuEfect(aplica un efect pe langa)

+++ src/bloc.cpp
@@ -9,7 +9,7 @@ Bloc::Bloc(float posX, float posY, float Width, float Height)
 
 Bloc::Bloc(float posX, float posY, float Width, float Height, const Caracter& owner)
     : Entitate(Width, Height, posX, posY), owner(&owner) {
-    lifespan = static_cast<int8_t>(owner.durabilitate_zid());
+    lifespan = owner.durabilitate_zid();
     dx_owner = (posX - owner.GetHitbox().x) / Entitate::get_factor_scalare();
 }
 
@@ -41,7 +41,7 @@ void Bloc::OnCollision(Entitate&) {
 }
 
 void Bloc::OnCollision(Sageata& s) {
-    lifespan -= static_cast<int8_t>(s.damage_perete());
+    lifespan -= s.damage_perete();
 }
 
 bool Bloc::TrebuieSters() const {
